refactor: 상환 시뮬레이션 service 인터페이스 구분 #72
